pyproject.toml
[project]
name = "problem-forge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.3.0",
    "pdfplumber>=0.11.6",
]
